Further refactoring of code to be modular
